Adding an Actor sample.
