+++ .gitignore
@@ -9,6 +9,13 @@
 /src/**/gen/**/*
 /src/**/StyleCop.Cache
 
+# components generated by Dapr when run in samples project directory.
+/samples/**/components/*.yaml
+
+# launchSettings.json for samples and tests
+/samples/**/Properties/launchSettings.json
+/test/**/Properties/launchSettings.json
+
 # MSBuild's log file
 msbuild.log
