+++ code.sln
@@ -31,6 +31,16 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Dapr.AspNetCore.Integration
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Dapr.AspNetCore.IntegrationTest.App", "test\Dapr.AspNetCore.IntegrationTest.App\Dapr.AspNetCore.IntegrationTest.App.csproj", "{342783B5-F75B-4752-A3E2-B8CB7D09C080}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Actor", "Actor", "{CA751B2C-E7F0-4AFA-BA5D-E91DD24E2847}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IDemoActor", "samples\Actor\IDemoActor\IDemoActor.csproj", "{234ADFE8-BA4D-4F75-BE77-5C32675FA3FA}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DemoActor", "samples\Actor\DemoActor\DemoActor.csproj", "{2A398D08-8708-4D97-B036-CCE415B6A715}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActorClient", "samples\Actor\ActorClient\ActorClient.csproj", "{9D0A3EB0-C47C-4C68-A9A2-B253EC81E6C4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "AspNetCore", "AspNetCore", "{DCBEAB71-0BDE-42F4-8187-F576942CB038}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -81,6 +91,18 @@ Global
 		{342783B5-F75B-4752-A3E2-B8CB7D09C080}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{342783B5-F75B-4752-A3E2-B8CB7D09C080}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{342783B5-F75B-4752-A3E2-B8CB7D09C080}.Release|Any CPU.Build.0 = Release|Any CPU
+		{234ADFE8-BA4D-4F75-BE77-5C32675FA3FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{234ADFE8-BA4D-4F75-BE77-5C32675FA3FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{234ADFE8-BA4D-4F75-BE77-5C32675FA3FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{234ADFE8-BA4D-4F75-BE77-5C32675FA3FA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2A398D08-8708-4D97-B036-CCE415B6A715}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2A398D08-8708-4D97-B036-CCE415B6A715}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2A398D08-8708-4D97-B036-CCE415B6A715}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2A398D08-8708-4D97-B036-CCE415B6A715}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9D0A3EB0-C47C-4C68-A9A2-B253EC81E6C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9D0A3EB0-C47C-4C68-A9A2-B253EC81E6C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9D0A3EB0-C47C-4C68-A9A2-B253EC81E6C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9D0A3EB0-C47C-4C68-A9A2-B253EC81E6C4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -91,12 +113,17 @@ Global
 		{B9C12532-0969-4DAC-A2F8-CA9208D7A901} = {27C5D71D-0721-4221-9286-B94AB07B58CF}
 		{62E41317-ED5D-4AA4-B129-C9E56C27354C} = {27C5D71D-0721-4221-9286-B94AB07B58CF}
 		{08D602F6-7C11-4653-B70B-B56333BF6FD2} = {27C5D71D-0721-4221-9286-B94AB07B58CF}
-		{2B015C86-32FB-4178-8F8A-26749B6FCE11} = {B2DB41EE-45F5-447B-95E8-38E1E8B70C4E}
-		{1FBDE4EF-D8B0-4BFE-B393-3CA4FE1944A2} = {B2DB41EE-45F5-447B-95E8-38E1E8B70C4E}
+		{2B015C86-32FB-4178-8F8A-26749B6FCE11} = {DCBEAB71-0BDE-42F4-8187-F576942CB038}
+		{1FBDE4EF-D8B0-4BFE-B393-3CA4FE1944A2} = {DCBEAB71-0BDE-42F4-8187-F576942CB038}
 		{383609C1-F43F-49EB-85E4-1964EE7F0F14} = {DD020B34-460F-455F-8D17-CF4A949F100B}
 		{B314AD5E-10AC-418A-B021-D4206BF37ACF} = {DD020B34-460F-455F-8D17-CF4A949F100B}
 		{0CD1912D-5E27-4A2A-A998-164792E0D006} = {DD020B34-460F-455F-8D17-CF4A949F100B}
 		{342783B5-F75B-4752-A3E2-B8CB7D09C080} = {DD020B34-460F-455F-8D17-CF4A949F100B}
+		{CA751B2C-E7F0-4AFA-BA5D-E91DD24E2847} = {B2DB41EE-45F5-447B-95E8-38E1E8B70C4E}
+		{234ADFE8-BA4D-4F75-BE77-5C32675FA3FA} = {CA751B2C-E7F0-4AFA-BA5D-E91DD24E2847}
+		{2A398D08-8708-4D97-B036-CCE415B6A715} = {CA751B2C-E7F0-4AFA-BA5D-E91DD24E2847}
+		{9D0A3EB0-C47C-4C68-A9A2-B253EC81E6C4} = {CA751B2C-E7F0-4AFA-BA5D-E91DD24E2847}
+		{DCBEAB71-0BDE-42F4-8187-F576942CB038} = {B2DB41EE-45F5-447B-95E8-38E1E8B70C4E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {65220BF2-EAE1-4CB2-AA58-EBE80768CB40}
