+++ samples/Actor/DemoActor/Program.cs
@@ -1,0 +1,40 @@
+﻿// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+// ------------------------------------------------------------
+
+namespace DaprDemoActor
+{
+    using Dapr.Actors.AspNetCore;
+    using Microsoft.AspNetCore;
+    using Microsoft.AspNetCore.Hosting;
+
+    /// <summary>
+    /// Class for host.
+    /// </summary>
+    public class Program
+    {
+        /// <summary>
+        /// Entyr point.
+        /// </summary>
+        /// <param name="args">Arguments.</param>
+        public static void Main(string[] args)
+        {
+            CreateWebHostBuilder(args).Build().Run();
+        }
+
+        /// <summary>
+        /// Creates a IWebHostBuilder.
+        /// </summary>
+        /// <param name="args">Arguments.</param>
+        /// <returns>IWebHostBuilder instance.</returns>
+        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
+            WebHost.CreateDefaultBuilder(args)
+                .UseStartup<Startup>()
+            .UseActors(actorRuntime =>
+            {
+                actorRuntime.RegisterActor<DemoActor>();
+            })
+            .UseUrls("http://localhost:5001/");
+    }
+}
