+++ samples/Actor/IDemoActor/IDemoActor.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
@@ -9,7 +9,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\Microsoft.Actions.Actors\Microsoft.Actions.Actors.csproj" />
+    <ProjectReference Include="..\..\..\src\Dapr.Actors\Dapr.Actors.csproj" />
   </ItemGroup>
 
 </Project>
