+++ samples/ControllerSample/Startup.cs
@@ -6,7 +6,6 @@
 namespace ControllerSample
 {
     using System.Text.Json;
-    using Dapr;
     using Microsoft.AspNetCore.Builder;
     using Microsoft.AspNetCore.Hosting;
     using Microsoft.Extensions.Configuration;
