--- test/ActionsActorDemo/ActionsActorDemo.sln
@@ -1,49 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.705
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActionsActorDemo", "ActionsActorDemo\ActionsActorDemo.csproj", "{21F32651-8A15-4100-855F-E8776EB25A35}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IDemoActor", "IDemoActor\IDemoActor.csproj", "{762D094B-7AE0-47F3-A124-54599091E4FE}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActorClient", "ActorClient\ActorClient.csproj", "{D8088875-1FE4-411C-AB2A-D832A0428A7E}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Actions.Actors", "..\..\src\Microsoft.Actions.Actors\Microsoft.Actions.Actors.csproj", "{72D81536-684B-47C3-ABAB-80F2928C35E9}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Actions.Actors.AspNetCore", "..\..\src\Microsoft.Actions.Actors.AspNetCore\Microsoft.Actions.Actors.AspNetCore.csproj", "{14D4F716-CD52-4849-B2F0-7F7147076B7E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{21F32651-8A15-4100-855F-E8776EB25A35}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{21F32651-8A15-4100-855F-E8776EB25A35}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{21F32651-8A15-4100-855F-E8776EB25A35}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{21F32651-8A15-4100-855F-E8776EB25A35}.Release|Any CPU.Build.0 = Release|Any CPU
-		{762D094B-7AE0-47F3-A124-54599091E4FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{762D094B-7AE0-47F3-A124-54599091E4FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{762D094B-7AE0-47F3-A124-54599091E4FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{762D094B-7AE0-47F3-A124-54599091E4FE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D8088875-1FE4-411C-AB2A-D832A0428A7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D8088875-1FE4-411C-AB2A-D832A0428A7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D8088875-1FE4-411C-AB2A-D832A0428A7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D8088875-1FE4-411C-AB2A-D832A0428A7E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{72D81536-684B-47C3-ABAB-80F2928C35E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{72D81536-684B-47C3-ABAB-80F2928C35E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{72D81536-684B-47C3-ABAB-80F2928C35E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{72D81536-684B-47C3-ABAB-80F2928C35E9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{14D4F716-CD52-4849-B2F0-7F7147076B7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{14D4F716-CD52-4849-B2F0-7F7147076B7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{14D4F716-CD52-4849-B2F0-7F7147076B7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{14D4F716-CD52-4849-B2F0-7F7147076B7E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {21CAD60C-0347-4170-894D-EDE5EEDA6564}
-	EndGlobalSection
-EndGlobal
