--- test/ActionsActorDemo/ActionsActorDemo/Properties/PublishProfiles/FolderProfile.pubxml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-This file is used by the publish/package process of your Web project. You can customize the behavior of this process
-by editing this MSBuild file. In order to learn more about this please visit https://go.microsoft.com/fwlink/?LinkID=208121. 
--->
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <WebPublishMethod>FileSystem</WebPublishMethod>
-    <PublishProvider>FileSystem</PublishProvider>
-    <LastUsedBuildConfiguration>Debug</LastUsedBuildConfiguration>
-    <LastUsedPlatform>Any CPU</LastUsedPlatform>
-    <SiteUrlToLaunchAfterPublish />
-    <LaunchSiteAfterPublish>True</LaunchSiteAfterPublish>
-    <ExcludeApp_Data>False</ExcludeApp_Data>
-    <ProjectGuid>21f32651-8a15-4100-855f-e8776eb25a35</ProjectGuid>
-    <publishUrl>bin\Debug\netcoreapp2.2\publish\</publishUrl>
-    <DeleteExistingFiles>False</DeleteExistingFiles>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <SelfContained>false</SelfContained>
-    <_IsPortable>true</_IsPortable>
-  </PropertyGroup>
-</Project>
